Added notification and lock tests to Hyperspace;  Fixed some bugs in the process.
git-svn-id: https://hypertable.googlecode.com/svn/trunk@280 01261837-d133-0410-8dcb-6792eb736156

+++ src/cc/Tools/hyperspace2/test/hyperspaceTest.cc
@@ -23,50 +23,64 @@
 #include <cstring>
 #include <vector>
 
-/**
-extern "C" {
-#include <poll.h>
-#include <signal.h>
-#include <sys/types.h>
-#include <unistd.h>
-}
-
-#include <boost/thread/thread.hpp>
-
 #include "Common/Error.h"
-#include "Common/FileUtils.h"
-#include "Common/TestHarness.h"
-#include "Common/StringExt.h"
-#include "Common/Usage.h"
-
-#include "AsyncComm/Comm.h"
-#include "AsyncComm/ConnectionManager.h"
-#include "AsyncComm/Event.h"
-#include "AsyncComm/ReactorFactory.h"
-
-#include "CommTestThreadFunction.h"
-
-*/
-
+#include "Common/InetAddr.h"
 #include "Common/Logger.h"
 #include "Common/ServerLauncher.h"
 #include "Common/System.h"
 
-using namespace hypertable;
+#include "AsyncComm/Comm.h"
+#include "AsyncComm/CommBuf.h"
+#include "AsyncComm/DispatchHandler.h"
+#include "AsyncComm/HeaderBuilder.h"
+#include "AsyncComm/ReactorFactory.h"
 
+
+using namespace hypertable;
 
 namespace {
   const char *usage[] = {
     "usage: hyperspaceTest",
     "",
-    "This program ...",
+    "This program tests Hyperspace using the hyperspace command interpreter.  It",
+    "Launches a Hyperspace server configured to use ./hyperspace as it's root",
+    "directory.  It then launches several hyperspace command interpreters and",
+    "issues commands to them, capturing the output for diffing.",
     0
   };
 
-  const int DEFAULT_PORT = 32998;
-  const char *DEFAULT_PORT_ARG = "--port=32998";
+  /**
+   *
+   */
+  class NotificationHandler : public DispatchHandler {
 
-  void IssueCommand(int fd, const char *command) {
+  public:
+
+    NotificationHandler() : DispatchHandler() { return; }
+
+    void SetPending() { mPending = true; }
+  
+    virtual void handle(EventPtr &eventPtr) {
+      boost::mutex::scoped_lock lock(mMutex);
+      mPending = false;
+      mCond.notify_all();
+    }
+
+    void WaitForNotification() {
+      boost::mutex::scoped_lock lock(mMutex);
+      while (mPending)
+	mCond.wait(lock);
+    }
+
+  private:
+    boost::mutex      mMutex;
+    boost::condition  mCond;
+    bool mPending;
+  };
+
+  NotificationHandler gNotifyHandler;
+
+  void IssueCommandNoWait(int fd, const char *command) {
     if (write(fd, command, strlen(command)) != strlen(command)) {
       perror("write");
       exit(1);
@@ -75,15 +89,20 @@ namespace {
       perror("write");
       exit(1);
     }
-    /**
-    std::string commandStr = (std::string)command + "\n";
-    if (write(fd, commandStr.c_str(), strlen(commandStr.c_str())) < 0) {
-      perror("write");
-      exit(1);
-    }
-    poll(0, 0, 100);
-    **/
   }
+
+  void IssueCommand(int fd, const char *command) {
+    gNotifyHandler.SetPending();
+    IssueCommandNoWait(fd, command);
+    gNotifyHandler.WaitForNotification();
+  }
+
+  int gFd1, gFd2, gFd3;
+
+  void OutputTestHeader(const char *name);
+  void BasicTest();
+  void NotificationTest();
+  void LockTest();
 
 }
 
@@ -91,9 +110,23 @@ namespace {
 int main(int argc, char **argv) {
   std::vector<const char *> masterArgs;
   std::vector<const char *> clientArgs;
-  int client1fd, client2fd, client3fd;
+  int error;
+  Comm *comm;
+  struct sockaddr_in addr;
 
   System::Initialize(argv[0]);
+  ReactorFactory::Initialize(1);
+
+  comm = new Comm();
+
+  if (!InetAddr::Initialize(&addr, "23451"))
+    exit(1);
+
+  if ((error = comm->CreateDatagramReceiveSocket(&addr, &gNotifyHandler)) != Error::OK) {
+    std::string str;
+    LOG_VA_ERROR("Problem creating UDP receive socket %s - %s", InetAddr::StringFormat(str, addr), Error::GetText(error));
+    exit(1);
+  }
 
   system("/bin/rm -rf ./hyperspace");
 
@@ -110,33 +143,26 @@ int main(int argc, char **argv) {
   clientArgs.push_back("hyperspace2");
   clientArgs.push_back("--config=./hyperspaceTest.cfg");
   clientArgs.push_back("--test-mode");
+  clientArgs.push_back("--notification-address=23451");
   clientArgs.push_back((const char *)0);
 
   {
     ServerLauncher master("../../Hyperspace/Hyperspace.Master", (char * const *)&masterArgs[0]);
     ServerLauncher client1("./hyperspace2", (char * const *)&clientArgs[0], "client1.out");
     ServerLauncher client2("./hyperspace2", (char * const *)&clientArgs[0], "client2.out");
+    ServerLauncher client3("./hyperspace2", (char * const *)&clientArgs[0], "client3.out");
 
-    client1fd = client1.GetWriteDescriptor();
-    client2fd = client2.GetWriteDescriptor();
+    gFd1 = client1.GetWriteDescriptor();
+    gFd2 = client2.GetWriteDescriptor();
+    gFd3 = client3.GetWriteDescriptor();
+    
+    BasicTest();
+    NotificationTest();
+    LockTest();
 
-    IssueCommand(client1fd, "mkdir foo");
-    //IssueCommand(client1fd, "open foo flags=READ|CREATE|WRITE event-mask=ATTR_SET|ATTR_DEL|CHILD_NODE_ADDED|CHILD_NODE_REMOVED|LOCK_ACQUIRED|LOCK_RELEASED");
-    IssueCommand(client1fd, "open foo flags=READ|CREATE|WRITE event-mask=ATTR_SET|ATTR_DEL|CHILD_NODE_ADDED|CHILD_NODE_REMOVED|LOCK_ACQUIRED|LOCK_RELEASED");
-    poll(0, 0, 2000);
-
-    IssueCommand(client2fd, "open foo flags=READ");
-    IssueCommand(client2fd, "mkdir foo/bar");
-    IssueCommand(client2fd, "delete foo/bar");
-    IssueCommand(client2fd, "attrset foo fox=\"Hello, World!\"");
-    IssueCommand(client2fd, "attrget foo fox");
-    IssueCommand(client2fd, "attrdel foo fox");
-    IssueCommand(client2fd, "attrget foo fox");
-
-    poll(0, 0, 1000);
-    IssueCommand(client2fd, "quit");
-    IssueCommand(client1fd, "quit");
-    poll(0, 0, 1000);
+    IssueCommand(gFd1, "quit");
+    IssueCommand(gFd2, "quit");
+    IssueCommand(gFd3, "quit");
   }
 
   if (system("diff ./client1.out ./client1.golden"))
@@ -145,72 +171,82 @@ int main(int argc, char **argv) {
   if (system("diff ./client2.out ./client2.golden"))
     return 1;
 
-  return 0;
-  
-
-#if 0
-
-  if (argc != 1)
-    Usage::DumpAndExit(usage);
-
-  srand(8876);
-
-  System::Initialize(argv[0]);
-  ReactorFactory::Initialize(1);
-
-
-
-
-  memset(&addr, 0, sizeof(struct sockaddr_in));
-  {
-    struct hostent *he = gethostbyname("localhost");
-    if (he == 0) {
-      herror("gethostbyname()");
-      return 1;
-    }
-    memcpy(&addr.sin_addr.s_addr, he->h_addr_list[0], sizeof(uint32_t));
-  }
-  addr.sin_family = AF_INET;
-  addr.sin_port = htons(DEFAULT_PORT);
-
-  comm = new Comm();
-
-  connManager = new ConnectionManager(comm);
-  connManager->Add(addr, 5, "testServer");
-  if (!connManager->WaitForConnection(addr, 30)) {
-    LOG_ERROR("Connect error");
+  if (system("diff ./client3.out ./client3.golden"))
     return 1;
-  }
-
-  CommTestThreadFunction threadFunc(comm, addr, "/usr/share/dict/words");
-
-  threadFunc.SetOutputFile("commTest.output.1");
-  thread1 = new boost::thread(threadFunc);
-
-  threadFunc.SetOutputFile("commTest.output.2");
-  thread2 = new boost::thread(threadFunc);
-
-  thread1->join();
-  thread2->join();
-
-  std::string tmpFile = (std::string)"/tmp/commTest" + (int)getpid();
-  std::string commandStr = (std::string)"head -" + (int)MAX_MESSAGES + " /usr/share/dict/words > " + tmpFile  + " ; diff " + tmpFile + " commTest.output.1";
-
-  if (system(commandStr.c_str()))
-    return 1;
-
-  commandStr = (std::string)"unlink " + tmpFile;
-  if (system(commandStr.c_str()))
-    return 1;
-
-  if (system("diff commTest.output.1 commTest.output.2"))
-    return 1;
-
-  delete connManager;
-  delete comm;
-  delete thread1;
-  delete thread2;
 
   return 0;
-#endif
+}
+
+
+namespace {
+
+  void OutputTestHeader(const char *name) {
+    std::string commandStr = (std::string)"echo " + name;
+    IssueCommand(gFd1, "echo");
+    IssueCommand(gFd1, commandStr.c_str());
+    IssueCommand(gFd2, "echo");
+    IssueCommand(gFd2, commandStr.c_str());
+    IssueCommand(gFd3, "echo");
+    IssueCommand(gFd3, commandStr.c_str());
+  }
+
+  void BasicTest() {
+    OutputTestHeader("<< BasicTest >>");
+    IssueCommand(gFd1, "mkdir dir1");
+    IssueCommand(gFd1, "mkdir how/now/brown/cow");
+    IssueCommand(gFd1, "delete foo");
+    IssueCommand(gFd1, "open foo flags=READ");
+    IssueCommand(gFd1, "open foo flags=READ|CREATE");
+    IssueCommand(gFd1, "open foo flags=READ|CREATE|EXCL");
+    IssueCommand(gFd1, "lock foo EXCLUSIVE");
+    IssueCommand(gFd1, "exists dir1");
+    IssueCommand(gFd1, "exists foo");
+    IssueCommand(gFd1, "exists bar");
+    IssueCommand(gFd1, "attrget foo testattr");
+    IssueCommand(gFd1, "delete dir1");
+    IssueCommand(gFd1, "close foo");    
+    IssueCommand(gFd1, "attrset foo testattr=\"Hello, World!\"");
+    IssueCommand(gFd1, "attrget foo testattr");
+    IssueCommand(gFd1, "attrdel foo testattr");
+  }
+
+  void NotificationTest() {
+    OutputTestHeader("<< NotificationTest >>");
+    IssueCommand(gFd1, "mkdir dir1");
+    IssueCommand(gFd1, "open dir1 flags=READ|CREATE|WRITE event-mask=ATTR_SET|ATTR_DEL|CHILD_NODE_ADDED|CHILD_NODE_REMOVED|LOCK_ACQUIRED|LOCK_RELEASED");
+
+    IssueCommand(gFd2, "mkdir dir1/foo");
+    IssueCommand(gFd2, "delete dir1/foo");
+    IssueCommand(gFd2, "open dir1 flags=READ|WRITE|LOCK");
+    IssueCommand(gFd2, "attrset dir1 fox=\"Hello, World!\"");
+    IssueCommand(gFd2, "attrget dir1 fox");
+    IssueCommand(gFd2, "attrdel dir1 fox");
+    IssueCommand(gFd2, "lock dir1 EXCLUSIVE");
+    IssueCommand(gFd2, "release dir1");
+
+    IssueCommand(gFd1, "close dir1");
+    IssueCommand(gFd2, "close dir1");
+    IssueCommand(gFd2, "delete dir1");
+  }
+
+  void LockTest() {
+    OutputTestHeader("<< LockTest >>");
+
+    IssueCommand(gFd1, "open lockfile flags=READ|CREATE|WRITE event-mask=ATTR_SET|ATTR_DEL|CHILD_NODE_ADDED|CHILD_NODE_REMOVED|LOCK_ACQUIRED|LOCK_RELEASED");
+    IssueCommand(gFd2, "open lockfile flags=READ|WRITE|LOCK");
+    IssueCommand(gFd3, "open lockfile flags=READ|WRITE|LOCK");
+    IssueCommand(gFd2, "lock lockfile EXCLUSIVE");
+    IssueCommandNoWait(gFd3, "lock lockfile EXCLUSIVE");
+    poll(0,0,1000);
+    IssueCommand(gFd2, "release lockfile");
+    poll(0,0,1000);
+    IssueCommand(gFd3, "release lockfile");
+    IssueCommand(gFd2, "lock lockfile SHARED");
+    IssueCommand(gFd3, "lock lockfile SHARED");
+    IssueCommand(gFd2, "release lockfile");
+    IssueCommand(gFd3, "release lockfile");
+    IssueCommand(gFd1, "close lockfile");
+    IssueCommand(gFd1, "delete lockfile");
+  }
+
 }
